# docs/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "StationPlay API",
    "description": "StationPlay's API: the server, its stations, their guide and streams, and things an Admin can have it do. Send an API token (made on StationPlay's Access tab) as `Authorization: Bearer <token>`: a Viewer token reads, an Admin token also does what's marked Admin. Version 1 only grows; see docs/api.md.",
    "version": "1",
    "license": {
      "name": "GPL-3.0-only"
    }
  },
  "paths": {
...
    "/api/v1/stations": {
      "get": {
        "summary": "The stations",
        "description": "Every station, with what's on now and next, and where to play it.",
        "operationId": "getStations",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Stations"
                }
              }
            }
          },
          "401": {
            "description": "No token, or one that isn't valid (while signing in is on)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "The token may not do this, or isn't accepted from there",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/guide": {
      "get": {
        "summary": "The guide",
        "description": "Each station's programs between two times: 6 hours from now unless asked; at most 2 days, starting within 7 days of now.",
        "operationId": "getGuide",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Where it starts (ms)",
              "title": "From"
            },
            "description": "Where it starts (ms)"
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Where it ends (ms)",
              "title": "To"
            },
            "description": "Where it ends (ms)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Guide"
                }
              }
            }
          },
          "400": {
            "description": "Something in the request wasn't understood",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "No token, or one that isn't valid (while signing in is on)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "The token may not do this, or isn't accepted from there",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Colors": {
        "properties": {
          "dark": {
            "type": "string",
            "title": "Dark",
            "description": "#RRGGBB"
          },
          "light": {
            "type": "string",
            "title": "Light",
            "description": "#RRGGBB"
          },
          "accent": {
            "type": "string",
            "title": "Accent",
            "description": "#RRGGBB"
          }
        },
        "type": "object",
        "required": [
          "dark",
          "light",
          "accent"
        ],
        "title": "Colors"
      },
      "Guide": {
        "properties": {
          "from": {
            "type": "integer",
            "title": "From",
            "description": "Where the guide starts (ms)"
          },
          "to": {
            "type": "integer",
            "title": "To",
            "description": "Where it ends (ms)"
          },
          "stations": {
            "items": {
              "$ref": "#/components/schemas/GuideStation"
            },
            "type": "array",
            "title": "Stations"
          }
        },
        "type": "object",
        "required": [
          "from",
          "to",
          "stations"
        ],
        "title": "Guide"
      },
...
      "GuideStation": {
        "properties": {
          "number": {
            "type": "integer",
            "title": "Number"
          },
          "programs": {
            "items": {
              "$ref": "#/components/schemas/Program"
            },
            "type": "array",
            "title": "Programs",
            "description": "Its programs in the time asked for, in order"
          }
        },
        "type": "object",
        "required": [
          "number",
          "programs"
        ],
        "title": "GuideStation"
      },
...
      "Program": {
        "properties": {
          "start": {
            "type": "integer",
            "title": "Start",
            "description": "When it starts, in milliseconds since 1970 (UTC)"
          },
          "end": {
            "type": "integer",
            "title": "End",
            "description": "When it ends (the commercials after it included)"
          },
          "kind": {
            "type": "string",
            "title": "Kind",
            "description": "episode or movie"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "The show's title, or the movie's"
          },
          "episodeTitle": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Episodetitle",
            "description": "The episode's title (null for a movie)"
          },
          "season": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Season"
          },
          "episode": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Episode"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          },
          "summary": {
            "type": "string",
            "title": "Summary",
            "description": "What it's about (may be empty)"
          },
          "art": {
            "type": "string",
            "title": "Art",
            "description": "Its picture, on the home network: the show's, or the movie's"
          },
          "special": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Special",
            "description": "What special it's part of: 'Feature Presentation', a block's name, or a marathon's"
          }
        },
        "type": "object",
        "required": [
          "start",
          "end",
          "kind",
          "title",
          "episodeTitle",
          "season",
          "episode",
          "year",
          "summary",
          "art",
          "special"
        ],
        "title": "Program"
      },
...
      "Station": {
        "properties": {
          "number": {
            "type": "integer",
            "title": "Number"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "logo": {
            "type": "string",
            "title": "Logo",
            "description": "Its logo (an address on the server)"
          },
          "colors": {
            "$ref": "#/components/schemas/Colors",
            "description": "Colors from its logo, as its Intro Bumper uses them"
          },
          "hls": {
            "type": "string",
            "title": "Hls",
            "description": "Where to play it (HLS): an address on the server. From the internet (with watching away from home on), an address of the token's own, which needs no token itself"
          },
          "now": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Program"
              },
              {
                "type": "null"
              }
            ],
            "description": "What's on now"
          },
          "next": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Program"
              },
              {
                "type": "null"
              }
            ],
            "description": "What's on next"
          }
        },
        "type": "object",
        "required": [
          "number",
          "name",
          "description",
          "logo",
          "colors",
          "hls",
          "now",
          "next"
        ],
        "title": "Station"
      },
      "Stations": {
        "properties": {
          "stations": {
            "items": {
              "$ref": "#/components/schemas/Station"
            },
            "type": "array",
            "title": "Stations",
            "description": "Every station, in number order"
          }
        },
        "type": "object",
        "required": [
          "stations"
        ],
        "title": "Stations"
      },
...
      "Error": {
        "type": "object",
        "title": "Error",
        "properties": {
          "detail": {
            "type": "string",
            "description": "Why, as a sentence to show"
          }
        },
        "required": [
          "detail"
        ]
      }
    },
    "securitySchemes": {
      "token": {
        "type": "http",
        "scheme": "bearer",
        "description": "An API token from the Access tab (or a StationPlay app's sign-in)"
      }
    }
  },
...
  "security": [
    {
      "token": []
    }
  ]
}